requirements.txt
aniso8601==10.0.0
annotated-types==0.7.0
attrs==25.1.0
blinker==1.9.0
bunnet==1.3.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
Flask-RESTful==0.3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy-model==0.2.0
MarkupSafe==3.0.2
mongoengine==0.29.1
motor==3.7.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11
python-dotenv==1.0.1
pytz==2025.1
referencing==0.36.2
rpds-py==0.23.1
six==1.17.0
toml==0.10.2
typing_extensions==4.12.2
Werkzeug==3.1.3
